requirements.txt
Flask==1.0.2
Jinja2==2.10
Werkzeug==0.14.1
mysql-connector==2.1.6

